requirements.txt
gradio>=5.0.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite
langchain-community
networkx>=3.0
ollama
pydantic>=2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
